fixed typos on the actual hardware

+++ hex2rgb.py
@@ -1,7 +1,7 @@
 def hex2rgb(value):
     value = value.lstrip('#')
     lv = len(value)
-    return tuple(int(value[i:i+lv/3], 16) for i in range(0, lv, lv/3))
+    return tuple(int(value[i:i+int(lv/3)], 16) for i in range(0, lv, int(lv/3)))
 
 def rgb2hex(rgb):
     return '%02x%02x%02x' % rgb
